requirements.txt
asgiref==3.6.0
Django==4.1.7
django-arrayfields==0.1.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
GDAL==3.6.3
numpy==1.24.2
Pillow==9.4.0
protobuf==4.21.12
pytz==2022.7.1
spatialite==0.0.3
sqlparse==0.4.3
